Progress towards keccak codegen
* Add //zirgen/Main:gen_zirgen that compiles .zir and splits out generated files that are split out instead of all in one big file
* Split out a bunch of zirgen driver.cpp to ease reuse by gen_zirgen
* Add zirgen support to build_circuit and bootstrap utility
* //zirgen/circuit filegroup target to build all circuits for faster `cargo bootstrap` build time
* Standardize --output-dir command line option for building circuits target directory
