+++ zirgen/Main/Main.h
@@ -1,0 +1,32 @@
+// Copyright 2024 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/Pass/PassManager.h"
+
+namespace zirgen {
+
+void registerZirgenCommon();
+
+void registerZirgenDialects(mlir::DialectRegistry& registry);
+
+void addAccumAndGlobalPasses(mlir::PassManager& pm);
+
+void addTypingPasses(mlir::PassManager& pm);
+
+mlir::LogicalResult checkDegreeExceeded(mlir::ModuleOp module, size_t maxDegree);
+
+} // namespace zirgen
