+++ zirgen/circuit/fib/BUILD.bazel
@@ -1,5 +1,9 @@
 load("//bazel/rules/zirgen:edsl-defs.bzl", "build_circuit")
 
+package(
+    default_visibility = ["//visibility:public"],
+)
+
 build_circuit(
     name = "fib",
     srcs = ["fib.cpp"],
